+++ test/errors/test_FPDF_errors.py
@@ -67,7 +67,7 @@ def test_adding_content_after_closing():
     pdf = fpdf.FPDF()
     pdf.set_font("helvetica", size=24)
     pdf.add_page()
-    pdf.cell(w=pdf.epw, h=10, txt="Hello fpdf2!", align="C")
+    pdf.cell(w=pdf.epw, txt="Hello fpdf2!", align="C")
     pdf.output()
     with pytest.raises(FPDFException) as error:
         pdf.add_page()
@@ -76,7 +76,7 @@ def test_adding_content_after_closing():
         == "A page cannot be added on a closed document, after calling output()"
     )
     with pytest.raises(FPDFException) as error:
-        pdf.cell(w=pdf.epw, h=10, txt="Hello again!", align="C")
+        pdf.cell(w=pdf.epw, txt="Hello again!", align="C")
     assert (
         str(error.value)
         == "Content cannot be added on a closed document, after calling output()"